+++ PieceOfCake.BlazorApp/Services/Interfaces/IMeasureUnitHttpService.cs
@@ -10,7 +10,7 @@ namespace PieceOfCake.BlazorApp.Services.Interfaces
     public interface IMeasureUnitHttpService
     {
         Task<Result<IEnumerable<MeasureUnitVm>>> GetAllMeasureUnits();
-        Task<Result<MeasureUnitVm>> GetMeasureUnitById(int measureUnitId);
+        Task<Result<MeasureUnitVm>> GetMeasureUnitById(long measureUnitId);
         Task<Result<MeasureUnitVm>> CreateMeasureUnit(MeasureUnitVm measureUnit);
         Task<Result<MeasureUnitVm>> UpdateMeasureUnit(MeasureUnitVm measureUnit);
         Task<Result> DeleteMeasureUnit(long measureUnitId);
